requirements.txt
aiogram>=3.15.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
rich>=13.0.0
textual>=0.79.0
